pyproject.toml
[tool.poetry]
name = "open-interpreter"
packages = [
    {include = "interpreter"},
]
version = "0.0.297"
description = "Ask GPT-4 to run code locally."
authors = ["Killian Lucas <killian@drinkwater.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.8"
rich = "^13.4.2"
tiktoken = "^0.4.0"
astor = "^0.8.1"
git-python = "^1.0.3"
tokentrim = "^0.1.2"
appdirs = "^1.4.4"
six = "^1.16.0"
ollama = "^0.2.0"
chromadb = "^0.4.24"
pyautogen = "^0.2.19"
pyautogui = "^0.9.54"
pynput = "^1.7.6"
pillow = "^10.3.0"
pytesseract = "^0.3.10"
SpeechRecognition = "^3.10.0"

# On non-windows systems, you can just `import readline`.
# On windows, `pyreadline3` replaces that, so you can also just `import readline`.
inquirer = "^3.1.3"
wget = "^3.2"
sounddevice = "^0.4.6"
numpy = "^1.26.4"
browser-use = "^0.1.18"
[tool.poetry.dependencies.pyreadline3]
version = "^3.4.1"
markers = "sys_platform == 'win32'"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
interpreter = "interpreter:cli"
unified-agent = "dev_framework.__main__:main"
